worked on district_test and district file and got one working test
