+++ test/district_test.rb
@@ -1,33 +1,14 @@
 require 'minitest/autorun'
 require 'minitest/pride'
+require './lib/district'
 
-class DistrictRepositoryTest < Minitest::Test
+class DistrictTest < Minitest::Test
 
-def test_that_load_data_returns_a_hash
-end
-
-def test_that_there_is_only_one_of_each_district
-  #could use #one? method asserting true
-end
-
-def test_that_search_by_name_returns_district_name
-  #upcased
-  #downcased
-  #mixed case
+def test_name_returns_upcase_string_name_of_district
+  d = District.new({:name => "Academy 20"})
+  assert_equal "ACADEMY 20", d.name
 end
 
 
-def test_that_search_of_wrong_name_returns_nil
-end
-
-def test_that_searching_by_name_fragment_returns_district_name
-  #upcased
-  #downcased
-  #mixed case
-
-end
-
-def test_that_searching_by_wrong_name_fragment_returns_empty_array
-end
 
 end
